fix: update icon config
There was a bug with the fallback values where the HOME icon was not
appearing at all. Using the `customize` attribute causes the design
tokens to be reset to `!00%`. The fix is to set hard values. Will
revisit if there are theme issues.

Changes to be committed:
modified:   src/auro-nav.js